requirements.txt
django==2.2
psycopg2-binary==2.8.4
django-model-utils==4.0.0
djangorestframework==3.11.0
djangorestframework_simplejwt==4.4.0
django-rest-swagger==2.2